Airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pandas>=1.3.0
yfinance==0.2.55